+++ project/__init__.py
@@ -7,6 +7,29 @@ from flask_bootstrap import Bootstrap
 from flask_elasticsearch import Elasticsearch
 from logging.handlers import RotatingFileHandler
 
+class SFNFormatter(logging.Formatter):
+    width = 45
+    datefmt='%Y-%m-%d %H:%M:%S'
+
+    def format(self, record):
+        cpath = f'{record.module}:{record.funcName}:[{record.lineno}]:{record.thread}'
+        cpath = cpath[-self.width:].ljust(self.width)
+        #record.message = record.getMessage()
+        levelName = f"[{record.levelname}]"
+        outputString = (f"{levelName:<10}: "
+                       f"{self.formatTime(record, self.datefmt)} : {cpath} : "
+                       f"{record.getMessage()}")
+
+        if record.exc_info:
+            # Cache the traceback text to avoid converting it multiple times
+            if not record.exc_text:
+                record.exc_text = self.formatException(record.exc_info)
+        if record.exc_text:
+            if outputString[-1:] != "\n":
+                outputString = outputString + "\n"
+            outputString = outputString + record.exc_text
+        return outputString
+
 
 # Initialize the app for Flask
 app = Flask(__name__)
@@ -19,7 +42,7 @@ app = Flask(__name__)
 # ---------- APPLICATION SETTINGS --------------
 #
 # Current version number of SafeNetworking
-app.config['VERSION'] = "2.0.3"
+app.config['VERSION'] = "2.1-dev"
 #
 # When set to True, this slows down the logging by only processing 1 event at a
 # time and allows us to see what is going on if there are bugs
@@ -35,10 +58,39 @@ app.config['SECRET_KEY'] = "\xfd{H\xe5<\x95\xf9\xe3\x96.5\xd1\x01O<!\xd5"
 app.config['BASE_DIR'] = os.path.abspath(os.path.dirname(__file__))
 #
 # Set the number of seconds for multi-threading to wait between processing calls
-app.config['POOL_TIME'] = 10
+app.config['DNS_POOL_TIME'] = 10
+app.config['URL_POOL_TIME'] = 10
+app.config['AF_POOL_TIME'] = 600
 #
-# Set the number of seconds - what the hell is this for?
-#app.config['SEC_PROCESS_POOL_TIME'] = 300
+# This is an internal flag that will probably never show up in the .panrc file
+# It is used to slow execution when it is True
+app.config['AF_POINTS_MODE'] = False
+#
+# Number of AF points left to slow down processing so we don't run out of points
+# When it reaches this point, it sets the AF_POINTS_MODE to True and it slows
+# execution to 1 event at a time.
+app.config['AF_POINTS_LOW'] = 5000
+#
+# Number of AF points left to stop processing all together
+app.config['AF_POINT_NOEXEC'] = 500
+#
+# Number of seconds to wait when AF_POINT_NOEXEC gets triggered.  This stops all
+# app execution and checks the AF points total at the specified interval.  When
+# the points total is higher than AF_POINT_NOEXEC it resumes execution.
+app.config['AF_NOEXEC_CKTIME'] = 3600
+#
+# Set the number of processes to run the DNS module.  This cannot be more than
+# 16 or it will kill the AF minute points.  The code will  take care of cases
+# where it is greater than 16 and this should only be adjusted down (never up).
+# Remember that DNS_POOL_COUNT and URL_POOL_COUNT have to share the total of 16.
+# The application will check and throw an error (but it will still run) if the
+# two settings together is more than 16.
+app.config['DNS_POOL_COUNT'] = 16
+#
+# Set the number of processes to run the URL module.  This cannot be more than
+# 16 or it will kill the AF minute points.  The code will  take care of cases
+# where it is greater than 16 and this should only be adjusted down (never up)
+app.config['URL_POOL_COUNT'] = 0
 #
 # When SafeNetworking is started, number of documents to read from the DB.  The
 # larger the number, the longer this will take to catch up.
@@ -104,6 +156,15 @@ app.config['ELASTICSTACK_VERSION'] = "6.1.1"
 #
 #
 #
+# ------------------------------ FLASK -----------------------------------------
+#
+# By default Flask listens to all ports - we will only listen to localhost
+# for security reasons, but keep the default port of 5000
+app.config['FLASK_HOST'] = "localhost"
+app.config['FLASK_PORT'] = 5000
+#
+#
+#
 # ----------------------------- MISCELLANEOUS ----------------------------------
 #
 app.config['AUTOFOCUS_API_KEY'] = "NOT-SET"
@@ -126,9 +187,9 @@ es = Elasticsearch(f"{app.config['ELASTICSEARCH_HOST']}:{app.config['ELASTICSEAR
 handler = RotatingFileHandler('log/sfn.log',
                             maxBytes=app.config['LOG_SIZE'],
                             backupCount=app.config['LOG_BACKUPS'])
-logFormat = logging.Formatter('%(asctime)s - %(module)s:%(funcName)s[%(lineno)i] - %(thread)d - [%(levelname)s] -- %(message)s')
+sfnFormatter = SFNFormatter()
 handler.setLevel(app.config["LOG_LEVEL"])
-handler.setFormatter(logFormat)
+handler.setFormatter(sfnFormatter)
 app.logger.addHandler(handler)
 app.logger.info(f"INIT - SafeNetworking application initializing with log level of {app.config['LOG_LEVEL']}")
 app.logger.info(f"ElasticSearch host is: {app.config['ELASTICSEARCH_HOST']}:{app.config['ELASTICSEARCH_PORT']}")